requirements.txt
pymupdf
sentence-transformers
chromadb
streamlit
spacy
python-dotenv
pandas
numpy

google-generativeai
fpdf
